pyproject.toml
[project]
name = "feuilles-annonces"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url~=2.2",
    "django~=5.2",
    "django-recurrence~=1.11",
    "django-solo~=2.4",
    "fpdf2~=2.8",
    "icalendar~=6.3",
]

[project.optional-dependencies]
server = ["gunicorn~=23.0"]
